// app/types/Sender.ts
export interface PostResponse {
    status: number;
    message: string;
    data: any;
}

export interface GetResponse {
    status: number;
    message: string;
    data: any;
}export interface PostResponse {
    status: number;
    message: string;
    data: any;
}

export interface GetResponse {
    status: number;
    message: string;
    data: any;
}